CI: Add CI job for VS2019 ARM64 builds
We need to write a Meson cross-compilation file on the fly here, hence the
additions in test-msvc.bat to set up the paths.

For obvious reasons, we can only run the build, not the tests, for the ARM64
Visual Studio CI job as the runner is an x86-based system; and like the 32-bit
VS2019 CI job this is only run manually or weekly.