CI: Add manual CI job for VS2019 ARM64 builds
This is currently only run manually, as we need to check whether the
subprojects are ready to be built as ARM64 binaries (for instance, libffi
needs to be quite recent to gain such support and the Meson patches for it
is probably not ready).

We need to write a Meson cross-compilation file on the fly here.

For obvious reasons, we can only run the build, not the tests, for the ARM64
Visual Studio CI job as the runner is an x86-based system.