gscanner: Fix types for read() and lseek() function calls
Use the right-width types for arguments and return values from these
calls, so we don’t get `-Wshorten-64-to-32` warnings.

The `read()` calls can never return enough to overflow an `int` anyway,
as the requested buffer size is limited.

The `lseek()` call could potentially have a huge offset to span, but
only if the `GScanner` was operating on a huge file and somehow had its
pointers at opposite ends of the file. I have not investigated to see if
that is possible.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3527