goption: Fix string length storage to use size_t rather than int
This moves all the string length handling in this part of `goption.c` to
use `size_t`. However, we need to assert that the string length is
at most `G_MAXINT` later on, as the length is passed to `printf()` to
add indentation, and it only accepts `int` for that kind of placeholder.

This introduces no functional changes, but does fix some
`-Wshorten-64-to-32` warnings.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #3527