GPortalNotificationBackend: Parse GBytesIcon and GLoadableIcon for v2
The notification portal v2 accepts a sealable `memfd`. We can load
`GBytesIcon` and `GLoadableIcon` into a `memfd` automatically. For version 1
of the portal we can load `GLoadableIcon` into a `GBytesIcon` to
increase support for more `GIcon` formats.

This doesn't include `GEmblemedIcon` it's likely that it will be
deprecated soon https://gitlab.gnome.org/GNOME/glib/-/issues/3544